Fixed Warnings from 64-Bit

+++ include/TMX/TmxMap.h
@@ -174,7 +174,7 @@ namespace Tmx
         const Tmx::Layer *GetLayer(int index) const { return layers.at(index); }
 
         /// Get the amount of layers.
-        int GetNumLayers() const { return layers.size(); }
+        int GetNumLayers() const { return (int)layers.size(); }
 
         /// Get the whole layers collection.
         const std::vector< Tmx::Layer* > &GetLayers() const { return layers; }
@@ -183,7 +183,7 @@ namespace Tmx
         const Tmx::TileLayer *GetTileLayer(int index) const { return tile_layers.at(index); }
 
         /// Get the amount of tile layers.
-        int GetNumTileLayers() const { return tile_layers.size(); }
+        int GetNumTileLayers() const { return (int)tile_layers.size(); }
 
         /// Get the whole collection of tile layers.
         const std::vector< Tmx::TileLayer* > &GetTileLayers() const { return tile_layers; }
@@ -192,7 +192,7 @@ namespace Tmx
         const Tmx::ObjectGroup *GetObjectGroup(int index) const { return object_groups.at(index); }
 
         /// Get the amount of object groups.
-        int GetNumObjectGroups() const { return object_groups.size(); }
+        int GetNumObjectGroups() const { return (int)object_groups.size(); }
 
         /// Get the whole collection of object groups.
         const std::vector< Tmx::ObjectGroup* > &GetObjectGroups() const { return object_groups; }
@@ -201,7 +201,7 @@ namespace Tmx
         const Tmx::ImageLayer *GetImageLayer(int index) const { return image_layers.at(index); }
 
         /// Get the amount of image layers.
-        int GetNumImageLayers() const { return image_layers.size(); }
+        int GetNumImageLayers() const { return (int)image_layers.size(); }
 
         /// Get the whole collection of image layers.
         const std::vector< Tmx::ImageLayer* > &GetImageLayers() const { return image_layers; }
@@ -216,7 +216,7 @@ namespace Tmx
         const Tmx::Tileset *GetTileset(int index) const { return tilesets.at(index); }
 
         /// Get the amount of tilesets.
-        int GetNumTilesets() const { return tilesets.size(); }
+        int GetNumTilesets() const { return (int)tilesets.size(); }
 
         /// Get the collection of tilesets.
         const std::vector< Tmx::Tileset* > &GetTilesets() const { return tilesets; }
