+++ include/TMX/TmxPolyline.h
@@ -52,7 +52,7 @@ namespace Tmx
         const Tmx::Point &GetPoint(int index) const { return points[index]; }
 
         /// Get the number of vertices.
-        int GetNumPoints() const { return points.size(); }
+        int GetNumPoints() const { return (int)points.size(); }
 
     private:
         std::vector< Tmx::Point > points;
