+++ include/TMX/TmxPropertySet.h
@@ -65,7 +65,7 @@ namespace Tmx
         Tmx::Color GetColorProperty(const std::string &name, Tmx::Color defaultValue = Tmx::Color()) const;
 
         /// Returns the amount of properties.
-        int GetSize() const { return properties.size(); }
+        int GetSize() const { return (int)properties.size(); }
 
         /// Checks if a property exists in the set.
         bool HasProperty( const std::string& name ) const;
